Update from task 2ef20695-12bd-4545-a973-6a45eae5d3af

+++ src/codegen/mod.rs
@@ -1,9 +1,10 @@
 use std::collections::HashMap;
+use rust_decimal::Decimal;
 use crate::parser::ast::{Program, Statement, Expression, Operator, BlockStatement};
 
 #[derive(Debug, Clone)]
 pub enum RuntimeValue {
-    Money { amount: f64, currency: String },
+    Money { amount: Decimal, currency: String },
     Boolean(bool),
 }
